requirements.txt
python-decouple
pytz
telethon